refactor: destroyArticle
