+++ src/actions/cms/articles/destroy.ts
@@ -3,23 +3,9 @@ import prisma from '@/prisma'
 import { createPrismaActionError } from '@/actions/error'
 import type { ActionReturn } from '@/actions/Types'
 import type { Article } from '@prisma/client'
+import { destroyArticle } from '@/server/cms/articles/destroy'
 
-export async function destroyArticle(id: number): Promise<ActionReturn<Article>> {
-    try {
-        const article = await prisma.article.delete({
-            where: { id }
-        })
-
-        // delete coverimage
-        await prisma.cmsImage.delete({
-            where: { id: article.coverImageId }
-        })
-
-        return {
-            success: true,
-            data: article
-        }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+export async function destroyArticleAction(id: number): Promise<ActionReturn<Article>> {
+    //TODO: auth
+    return await destroyArticle(id)
 }
